run-stage: access externally provided ROMs
This allows access to ROM modules that are provided from outside
of the project.

The feature is activated as in the code snippet below:

<requires>
    <rom label="name_of_rom"/>
</requires>

The following steps enable you to use this feature:
  - create var/roms in your project
  - place the files to access in this directory
  - add the <rom label=""/> nodes to your runtime

The Linux run-stage creates a symbolic link to map the directory.

examples/external_roms shows how to use this feature.

Issue #55